move creation of exchanges to main.rs
